fix #68 calculation of diff with previous value not correct if readings less than 5 minutes apart
Calculation of diff was taking the slope and then multiplied with 5 minutes. This seems not correct (strange nobody every noticed ?)

Probably because originally this was written for Dexcom transmitters, which give a reading every 5 minutes.

Although there too the diff is not correct if time difference > 5 minutes (10 minutes, 15, ...)

This changes will give another diff value in home screen, speak reading, watch, notification, text in alert.

Triggering of alerts is not impacted.