fix #86 fix for oop web if libreDerivedAlgorithmParameters are not correct
There seem to be cases where libreDerivedAlgorithmParameters are not correctly downloaded from the oop web server.
In that case those wrong values are stored and the sensor will never give a calculated value.

This is detected by checking that libreDerivedAlgorithmParameters.slope_slope != 0.0
If it iss, the parameters will be redownloaded from the oop web server