Volume tests.
Two volume tests are possible in the settings screen :
- volume of sound used for alarms where override mute = on
When clicking the setting, the xdrip sound is played. The volume can be turned up or down while the sound is playing, with the device volume buttons
- volume of sound used for alarms where override mute = off AND for missed reading alarms.
When clicking the setting, the xDrip sound is played. When using one of the volume buttons, the sound immediately stops playing. The volume can be changed but the user will not hear the current volume. A new test needs to be intiated to hear the new volume.

Note : missed reading alerts will actually never be able to override mute. Even though an alarm type with override mute = on
That's always been the case, probably not known because I never mentioned it