changed maxWaitForpacketInSeconds for MiaoMiao to 5 seconds. This should give less missed readings.
What happened if it's longer (it was 60 seconds before): let's say first packet is received (starting with 28), then immediately after that a disconnect occurs, immediately followed by a reconnect. In this case a new reading will be sent, ie the next packet will actually be again the first packet starting with 28.
But, as this is much less than 60 seconds after having received the very first packet, the rxBuffer isn't reset. So this packet is added as the second packet, and then the second packet is added as the third. As a result, the state 'Buffer complete' is reached too soon. Resulting in a missed reading