fix #189 Libre 2 with Bubble : when using xDrip algo, app asks for new calibration randomly
There was a wrong calculation of sensorId. A first connect, patchInfo was not yet available, resulting in sensorId starting with 0. Next reading, the patchInfo is available, resulting in sensorId starting with 3. Which results in a new sensor being detected, and a stop/start of the sensor, and so a new calibration request.