fix #201 - relaunches the app after 'App removed from memory' or 'App crashed'
create centralManager with a CBCentralManagerOptionRestoreIdentifierKey. This to ensure that iOS relaunches the app whenever it's killed either due to a crash or due to lack of memory
iOS will restart the app as soon as a bluetooth transmitter tries to connect (which is under normal circumtances immediately
the function willRestoreState is an empty function and that seems to be enough to make it work
see https://developer.apple.com/library/archive/qa/qa1962/_index.html

It's a quick solution, it will still miss one reading, because it doesn't use the data at the moment it's relaunched. (function willRestoreState is an empty function)