if willRestorState is called, then make sure a trace is done, so we can see in the issue reports that this has happened
in BluetoothTransmitter : centralManager is initialized with key CBCentralManagerOptionRestoreIdentifierKey. The assigned value must always be the same (per bluetooth peripheral), otherwise willRestoreState doesn't get called.

Previously it was always a random value, ie always a different value.

Now it's always the same value, consisting of the appname + the mac address of the peripheral, meaning unique per peripheral but the same value for each app start.

One exception is the first scan, because then we don't know yet the mac address, in that case a random value is used.