fix #205 : Watch and notification : option to specify minimum time interval between two events or two notifications.
This is not completely as requested in the issue https://github.com/JohanDegraeve/xdripswift/issues/205

It does allow watch/Libre 2 users to have a new reading every minute.
It also allows all users to have specify the minimum time between to calendar events on the watch, or to have a notification with a reading.

For the notification : an overrule of the interval is done if there's been a dis/reconnect, which means if there was a disconnect/reconnect, the next reading will generate a notification, no matter the value of the interval.

Default value for the interval = 0