fix bug, no upload new readings after Internet failure
The bug was that after restore of failed Internet connection, new readings are not uploaded to NS.

- as mentioned in the comments of func uploadData(dataToUpload: Any, httpMethod: String?, path: String, completionHandler: (() -> ())?)
completionhandler should be called only when upload was successful.
But the result of the upload was never checked. So probably, in some error cases, completionhandler was called with success. As a result, the upload was considered as successful and not redone when internet comes back

- also some updates in func uploadDataAndGetResponse. Set nightScoutResult to .failed when there was an error.