call cgmTransmitterInfoChanged only if blePeripheral is the active cgm transmitter
it's possible to create multiple cgm transmitters (eg a Libre 2 and a dexcom) but only one can be active, ie connected.
But one could go to the settings for the non-active CGM, and change the setting 'Use Transmitter algorithm'.
In that case, the setting was changed also for the active CGM, causing incorrect status.

this is fixed now
haha