initial commit calibration assistant
This commit introduces a calibration assistant function to xdripswift.
When the user clicks calibrate, the previous x minutes of BG data is analysed (delta change, std deviation and current BG level), weightings are applied to the deviations calculated from the programmed limits and a result is output. Based upon this result to user is allowed to calibrate normally or they are asked to consider waiting, or they are told directly to not calibrate.
A debug option is in developer settings to show the calculations in the UIAlert to allow troubleshooting.
The results and recommendation are also logged to the trace file.
The user can enable a visual assistant that will change the calibration icon colour (every time a new BG is processed) to indicate if it is a good time to calibrate or not.
EN and ES translations are included