Ensure treatments sync when app comes to foreground
Schedule a call to updateLabelsAndChart when the app comes to the foreground, with a delay of 0.5 seconds. Because the application state is not immediately to .active, as a result, updates may not happen - especially the synctreatments may not happen because this may depend on the application state - by making a call just half a second later, when the status is surely = .active, the treatments sync will be done.

This is specifically important in case NightScout is enabled but with an upload schema that is currently not active. If the app is in foreground, then treatment sync must happen. And this immediately after the app is coming to the foreground