in LoopManager, check that lastReadings.first is not nil
The app was suddenly crashing each time there was a new reading. Apparently, even though lastReadings.count is not 0, still lastReadings.first was nil. But there is a first unwrapping done of lastReadings.first after this check, causing a crash.
Now a check is done to see if lastReadings.first is not nil before proceeding.