Added a 'cover' to hide the main screen when the app is sent to the background.
Some apps utilise this kind of feature to prevent others from viewing a screen with personal data displayed, but it also (in this case, I hope) has the added bonus of making sure that an out-of-date BG reading isn't displayed when the App Switcher screen is shown on the device. The aim is to force the user to bring the app front and centre and not to rely on any cached frame grab.