This update gives xDrip an Ambulatory Glucose Profile
http://www.agpreport.org/agp/about
https://diabetes.medicinematters.com/blood-glucose-monitoring/continuous-glucose-monitoring/ambulatory-glucose-profile-what-why-and-how/16063922

It's accessed by a long press on the statistics view. There are 4 buttons along the bottom of the AGP view: 7 days, 14 days, 30 days and 90 days. The white chart line is the 50th quartile, the light area surrounding it is the 25th - 75th quartile and the darker area surrounding that, is the 10th - 90th quartile.

If the user touches on the chart, a timebar will appear with the glucose quartile figures for that point in an 'average' day - note there is also a 'm' figure which is the symbol for average (not to be confused with the 50th quartile). This 'm' is as international a symbol as I can find for the mathematical average. As with the buttons icons, I've sought not to use words which would need to be translated and I'm hoping that a number followed by the calendar symbol will indicate that it's 'x days'. This falls a little bit apart with the Q's for the quartiles and the views labelling.

There are 24 sections to the chart from midnight to 23:59, each covering an hour in a day. The timebar that appears has the hour of the day that the figures refer to.

The chart is dismissed by the 'X' button in the top right of the view.

During the course of this feature addition, I've used a UniversalBGLevel struct. I hope others will also find this useful as it's an object that can hold a reading in both mg/dl and mmol/l and provides convenience functions for getting unitised strings without the client code having to query the UserDefaults for which glucose measurement unit the user has picked.

There are also extensions to the Swift Double numeric type which provides MMOL and MGDL primitives - hopefully they can be of use too as it enables some mathematical operations to levels directly and makes for easier to understand code with quick help showing which unit type a BG iVar is.