add StandBy mode to the .systemSmall iOS widget
Will detect if the .systemSmall widget is being used in the StandBy mode and provide a simplified view with bigger text etc.

If the current time is between 22hrs and 08hrs, then we will *also* force a high contrast view so that it renders nicely in the StandBy Night Mode. This can be disabled in the developer settings. It is enabled by default.