Add switchable OS-AID app group sharing
Renames (in the UI only) "Loop Share" to "OS-AID Share"

Also creates an additional shared app group for Trio to help with this Issue: https://github.com/nightscout/Trio/issues/313

group.org.nightscout.${DEVELOPMENT_TEAM}.trio.trio-app-group

In Developer Settings, instead of "Suppress Loop Share", the user can now chose to disable share or choose to set it to the usual "LoopKit" app group, or the new "Trio" app group.

All other userdefault data is always stored in the usual LoopKit app group.

The share types are handled by a new LoopShareType enum that manages the UI strings and app group suite names in an opaque way to the rest of the app. It makes it easy to add further examples as required with no further changes to the app itself.