Rename "Transmitter Algorithm" to "Native Algorithm"
This makes it clearer... the idea of a transmitter algorithm is confusing for Libre users and doesn't really explain what it is.

At least for now, Native explains that the values will resemble (or be identical to) the factory algorithms in all cases. This may change in the future, but for now, it's accurate enough for the description.