change LibreLinkUpFollower to work with no active sensor data
Since LLU 4.12.0, it seems that the server response can now sometimes return an empty activeSensor attribute. This caused xDrip4iOS to throw an error and not process downloaded data.

This issue is known to affect just a handful of users and does not always happen. Sometimes the server responds correctly with all data.

We'll now just treat this as optional and ignore it (and adapt/hide the UI as needed) if necessary.