add check for keep-alive to Contact Image
This will display "OFF" in the contact image if the user is in follower mode with background keep-alive disabled, or if they disable the contact image function.

Also change settings text from "create contact" to "enable contact image"

Add warning text to "Enable Contact Image" UISwitch if the user is in follower mode with keep-alive disabled.