change to main chart hours selection
Change from 3/5/12/24 hour chart widths to 3/5/8/12 chart widths.

The old 24 hour option made no sense for a small chart (and the phone can be rotated to give a full 24 hour view)

Note that although the UI showed 6hrs, it actually used 5hrs for the last 4 years and nobody noticed the missing hour. So we'll leave 5hrs as the default value and change the UI to reflect the reality.