Live Activity restart App Intent (Shortcuts)
Adds an app intent (LiveActivityIntent) so that an automation in the Shortcuts app can restart an ongoing Live Activity.

Usually best done by adding the automation to run during the night to prevent the Live Activity from being cancelled if the app isn't opened in 8 hours.

Could also be set to run every 6 hours (as an example) to keep the Live Activity running forever without needing to open the app at all.

Co-Authored-By: dugamarian <35396265+dugamarian@users.noreply.github.com>