build_xdrip.yml: Disable keepalive action
The gautamkrishnar/keepalive-workflow has been taken down by GitHub, and GitHub Actions builds are failing due to this. 

Disabling the keepalive-workflow because the repo is no longer available. A permanent fix will need to be implemented later, this is a bandaid to prevent all build workflows to fail.

This is also the same change that was made for LoopWorkspace:
LoopKit/LoopWorkspace#248