add "main chart auto-reset" option to the main chart settings
this will automatically rescale and reset the main chart back to current glucose values and current time whenever the app  comes to the foreground. This is useful if the user had left the chart scrolled back in time, or if the y-axis had been extended due to an older high glucose value.