Add the possibility to ignore FollowerDataSourceTypes via override file
Add to xDripConfigOverride file if needed using an array of the rawValues of the types to be hidden/ignored.

case nightscout = 0
case libreLinkUp = 1
case libreLinkUpRussia = 2
case dexcomShare = 3

For example to hide/ignore Dexcom Share Follower:

IGNORE_FOLLOWER_TYPES=[3]

Note: Nightscout cannot be disabled or hidden. It will always be available and is the fallback option if a previously stored FollowerDataSourceTypes is now ignored.